Change the load_in_register functions to accomodate other tests
some operations don't write on the registers but intead on the memory
address. Since the destination memory address is not passed I needed a
way to "communicate" that the result is in the memory address decided by
the test.
I decide to use a pointer for this purpose.
All load tests still compile. This commit should not affect
functionality of previous features

Signed-off-by: Roberto Bertelli <robbybertelli@gmail.com>